# Implement world-local component id resolution and complete rewrite of type registration code