--- a/src/Flecs.NET.Benchmarks/Flecs.NET.Benchmarks.csproj
+++ b/src/Flecs.NET.Benchmarks/Flecs.NET.Benchmarks.csproj
@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFrameworks>netcoreapp3.1;net6.0;net7.0</TargetFrameworks>
+        <TargetFrameworks>netcoreapp3.1;net6.0;net7.0;net8.0</TargetFrameworks>
         <Nullable>enable</Nullable>
         <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
         <NoWarn>$(NoWarn);CS8618</NoWarn>
@@ -11,7 +11,7 @@
     <Import Project="../Flecs.NET.Native/Flecs.NET.Native.targets"/>
 
     <ItemGroup>
-        <PackageReference Include="BenchmarkDotNet" Version="0.13.7"/>
+        <PackageReference Include="BenchmarkDotNet" Version="0.13.12"/>
         <ProjectReference Include="../Flecs.NET/Flecs.NET.csproj"/>
     </ItemGroup>