--- a/src/Flecs.NET.Bindings/Flecs.g.cs
+++ b/src/Flecs.NET.Bindings/Flecs.g.cs
@@ -13127,7 +13127,9 @@ namespace Flecs.NET.Bindings
         {
             public static readonly System.Collections.Generic.List<string> DllFilePaths;
 
-            public static System.IntPtr _libraryHandle = System.IntPtr.Zero;
+            public static System.IntPtr LibraryHandle = System.IntPtr.Zero;
+
+            public static readonly object Lock = new object ();
 
             public static bool IsLinux => System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Linux);
 
@@ -13184,12 +13186,12 @@ namespace Flecs.NET.Bindings
             public static System.IntPtr GetExport(string symbol)
             {
 #if NET5_0_OR_GREATER
-            return System.Runtime.InteropServices.NativeLibrary.GetExport(_libraryHandle, symbol);
+            return System.Runtime.InteropServices.NativeLibrary.GetExport(LibraryHandle, symbol);
 #else
                 if (IsLinux)
                 {
                     GetLastErrorLinux();
-                    System.IntPtr handle = GetExportLinux(_libraryHandle, symbol);
+                    System.IntPtr handle = GetExportLinux(LibraryHandle, symbol);
                     if (handle != System.IntPtr.Zero)
                         return handle;
                     byte* errorResult = GetLastErrorLinux();
@@ -13202,7 +13204,7 @@ namespace Flecs.NET.Bindings
                 if (IsOsx)
                 {
                     GetLastErrorOsx();
-                    System.IntPtr handle = GetExportOsx(_libraryHandle, symbol);
+                    System.IntPtr handle = GetExportOsx(LibraryHandle, symbol);
                     if (handle != System.IntPtr.Zero)
                         return handle;
                     byte* errorResult = GetLastErrorOsx();
@@ -13214,7 +13216,7 @@ namespace Flecs.NET.Bindings
 
                 if (IsWindows)
                 {
-                    System.IntPtr handle = GetExportWindows(_libraryHandle, symbol);
+                    System.IntPtr handle = GetExportWindows(LibraryHandle, symbol);
                     if (handle != System.IntPtr.Zero)
                         return handle;
                     int errorCode = GetLastErrorWindows();
@@ -13237,45 +13239,51 @@ namespace Flecs.NET.Bindings
                     fileExtension = ".dll";
                 else
                     throw new System.InvalidOperationException("Can't determine native library file extension for the current system.");
+                System.IntPtr handle = default;
                 foreach (string dllFilePath in DllFilePaths)
                 {
                     string fileName = System.IO.Path.GetFileName(dllFilePath);
                     string parentDir = $"{dllFilePath}/..";
                     string searchDir = System.IO.Path.IsPathRooted(dllFilePath) ? System.IO.Path.GetFullPath(parentDir) + "/" : System.IO.Path.GetFullPath(System.AppDomain.CurrentDomain.BaseDirectory + parentDir) + "/";
-                    if (TryLoad($"{searchDir}{fileName}", out _libraryHandle))
-                        return;
-                    if (TryLoad($"{searchDir}{fileName}{fileExtension}", out _libraryHandle))
-                        return;
-                    if (TryLoad($"{searchDir}lib{fileName}", out _libraryHandle))
-                        return;
-                    if (TryLoad($"{searchDir}lib{fileName}{fileExtension}", out _libraryHandle))
-                        return;
+                    if (TryLoad($"{searchDir}{fileName}", out handle))
+                        goto Return;
+                    if (TryLoad($"{searchDir}{fileName}{fileExtension}", out handle))
+                        goto Return;
+                    if (TryLoad($"{searchDir}lib{fileName}", out handle))
+                        goto Return;
+                    if (TryLoad($"{searchDir}lib{fileName}{fileExtension}", out handle))
+                        goto Return;
                     if (!fileName.StartsWith("lib") || fileName == "lib")
                         continue;
                     string unprefixed = fileName.Substring(4);
-                    if (TryLoad($"{searchDir}{unprefixed}", out _libraryHandle))
-                        return;
-                    if (TryLoad($"{searchDir}{unprefixed}{fileExtension}", out _libraryHandle))
-                        return;
+                    if (TryLoad($"{searchDir}{unprefixed}", out handle))
+                        goto Return;
+                    if (TryLoad($"{searchDir}{unprefixed}{fileExtension}", out handle))
+                        goto Return;
                 }
 
 #if NET7_0_OR_GREATER
-                _libraryHandle = System.Runtime.InteropServices.NativeLibrary.GetMainProgramHandle();
+                handle = System.Runtime.InteropServices.NativeLibrary.GetMainProgramHandle();
 #else
                 if (IsLinux)
-                    _libraryHandle = LoadLibraryLinux(null, 0x101);
+                    handle = LoadLibraryLinux(null, 0x101);
                 else if (IsOsx)
-                    _libraryHandle = LoadLibraryOsx(null, 0x101);
+                    handle = LoadLibraryOsx(null, 0x101);
                 else if (IsWindows)
-                    _libraryHandle = GetModuleHandle(null);
+                    handle = GetModuleHandle(null);
 #endif
+                Return:
+                    LibraryHandle = handle;
             }
 
             public static void* LoadDllSymbol(string variableSymbol, out void* field)
             {
-                if (_libraryHandle == System.IntPtr.Zero)
-                    ResolveLibrary();
-                return field = (void*)GetExport(variableSymbol);
+                lock (Lock)
+                {
+                    if (LibraryHandle == System.IntPtr.Zero)
+                        ResolveLibrary();
+                    return field = (void*)GetExport(variableSymbol);
+                }
             }
         }
     }