--- a/src/Flecs.NET.Tests/CSharp/Core/TypeRegistrationTests.cs
+++ b/src/Flecs.NET.Tests/CSharp/Core/TypeRegistrationTests.cs
@@ -3,60 +3,34 @@ using Xunit;
 
 namespace Flecs.NET.Tests.CSharp.Core
 {
-    public unsafe class TypeRegistrationTests
+    public unsafe class TypeRegistrationTests // TODO: Add type registration tests
     {
         [Fact]
-        private void TypeClassComponentSize()
+        private void StructSize()
         {
             using World world = World.Create();
-
-            Type<Position>.Id(world);
             Assert.Equal(sizeof(Position), Type<Position>.Size);
         }
 
         [Fact]
-        private void TypeClassTagSize()
+        private void EmptyStructSize()
         {
             using World world = World.Create();
-
-            Type<Tag>.Id(world);
             Assert.Equal(0, Type<Tag>.Size);
         }
 
         [Fact]
-        private void TypeClassEnumSize()
+        private void EnumSize()
         {
             using World world = World.Create();
-
-            Type<StandardEnum>.Id(world);
-            Assert.Equal(sizeof(StandardEnum), Type<StandardEnum>.Size);
+            Assert.Equal(sizeof(Color), Type<Color>.Size);
         }
 
         [Fact]
-        private void ComponentFactoryComponentSize()
+        private void ClassSize()
         {
             using World world = World.Create();
-
-            world.Component<Position>();
-            Assert.Equal(sizeof(Position), Type<Position>.Size);
-        }
-
-        [Fact]
-        private void ComponentFactoryTagSize()
-        {
-            using World world = World.Create();
-
-            world.Component<Tag>();
-            Assert.Equal(0, Type<Tag>.Size);
-        }
-
-        [Fact]
-        private void ComponentFactoryEnumSize()
-        {
-            using World world = World.Create();
-
-            world.Component<StandardEnum>();
-            Assert.Equal(sizeof(StandardEnum), Type<StandardEnum>.Size);
+            Assert.Equal(sizeof(string), Type<string>.Size);
         }
     }
 }