--- a/src/Flecs.NET.Tests/Cpp/ModuleTests.cs
+++ b/src/Flecs.NET.Tests/Cpp/ModuleTests.cs
@@ -8,11 +8,6 @@ namespace Flecs.NET.Tests.Cpp
 {
     public class ModuleTests
     {
-        public ModuleTests()
-        {
-            FlecsInternal.Reset();
-        }
-
         [Fact]
         private void Import()
         {