This pull request refactors the component registration/lookup behavior to use a [world-local array lookup](https://github.com/SanderMertens/flecs/pull/1036) instead of a static field lookup when attempting to resolve component ids. This approach addresses a [known issue](https://github.com/SanderMertens/flecs/issues/1032) in multi-world applications where registering components in a different order can cause failed assertions due to id clashing.

Each type is given a unique index that denotes its position in the world-local array. The id for a type will be found at the same array index for all worlds.
```cs
public static class Type<T> {
    public static int TypeIndex = Interlocked.Increment(ref Ecs.CacheIndexCount);
}

public struct World {
    public List<ecs_entity_t> Cache; // world.Cache[Type<Position>.TypeIndex]
}
```
```cs
using World a = World.Create();
a.Component<Position>(); // Type index 1 for all worlds
a.Component<Velocity>(); // Type index 2 for all worlds

using World b = World.Create();
b.Component<Mass>(); // Type index 3 for all worlds
b.Component<Position>(); // Type index 1 still
b.Component<Velocity>(); // Type index 2 still
```
The following is done to determine whether or not a component needs to be registered for a particular world:
- If the type index is not within the bounds of the world-local array, it is not registered. Grow the array and register the component id at the given type index.
```cs
if (index >= world.Cache.Count)
    // Grow array and register id at cache[index].
```
- If the type index is within the bounds of the world-local array and the component id occupying its location is 0, it is not registered. Register the component id.
```cs
if (world.Cache[index] == 0)
    // Register id at cache[index].
```
___
The benchmark results below show decent performance improvements compared to the previous implementation.

**Before: (Static Field)**
```Console
| Method  | EntityCount | Mean     | Error     | StdDev    |
|-------- |------------ |---------:|----------:|----------:|
| TypeTag | 100000      | 8.399 ms | 0.0258 ms | 0.0216 ms |
| IntTag  | 100000      | 4.866 ms | 0.0585 ms | 0.0547 ms |
```

**After:  (World-Local Array)**
```Console
| Method  | EntityCount | Mean     | Error     | StdDev    |
|-------- |------------ |---------:|----------:|----------:|
| TypeTag | 100000      | 7.150 ms | 0.0334 ms | 0.0296 ms |
| IntTag  | 100000      | 4.861 ms | 0.0365 ms | 0.0323 ms |
```