@BeanCheeseBurrito 2024-04-18
Further performance improvements from removing unnecessary string argument from ``Type<T>.Id()`` function.
``Type<T>.Id(ecs_world_t*, string?)`` -> ``Type<T>.Id(ecs_world_t*)``

```Console
| Method  | EntityCount | Mean     | Error     | StdDev    |
|-------- |------------ |---------:|----------:|----------:|
| TypeTag | 100000      | 5.456 ms | 0.0339 ms | 0.0300 ms |
| IntTag  | 100000      | 4.723 ms | 0.0319 ms | 0.0298 ms |
```