@BeanCheeseBurrito 2024-04-19
Adding aggressive inlining to id lookup functions and adding a [SuppressGCTransition](https://learn.microsoft.com/en-us/dotnet/api/system.runtime.interopservices.suppressgctransitionattribute?view=net-8.0) attribute to the ``ecs_get_binding_ctx`` function gives even more performance improvements.

```Console
| Method              | EntityCount | Mean     | Error     | StdDev    |
|-------------------- |------------ |---------:|----------:|----------:|
| AddRemoveTypeTag    | 100000      | 5.128 ms | 0.0195 ms | 0.0163 ms |
| AddRemoveIntegerTag | 100000      | 4.722 ms | 0.0756 ms | 0.0670 ms |
```